+++ src/models/index.ts
@@ -2,3 +2,7 @@ export * from './enums';
 export * from './search';
 export * from './item';
 export * from './profile';
+export * from './review';
+export * from './related';
+export * from './shops';
+export * from './social';